requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
opencv-python
torch==2.5.1
torchvision==0.20.1
diffusers
transformers
accelerate
safetensors
ftfy
basicsr
realesrgan
gfpgan
facexlib
replicate
requests
python-dotenv
xformers
streamlit-image-select